+++ bindings/nostr-sdk-ffi/python/examples/tor.py
@@ -1,5 +1,5 @@
 import asyncio
-from nostr_sdk import Keys, Client, Options, EventBuilder, Connection, ConnectionTarget, init_logger, LogLevel
+from nostr_sdk import Keys, Client, ClientBuilder, Options, EventBuilder, Connection, ConnectionTarget, init_logger, LogLevel
 from datetime import timedelta
 
 
@@ -12,7 +12,7 @@ async def main():
     # Configure client to use embedded tor for `.onion` relays
     connection = Connection().embedded_tor().target(ConnectionTarget.ONION)
     opts = Options().connection(connection).connection_timeout(timedelta(seconds=60))
-    client = Client.with_opts(keys, opts)
+    client = ClientBuilder().signer(keys).opts(opts).build()
 
     await client.add_relay("wss://relay.damus.io")
     await client.add_relay("ws://oxtrdevav64z64yb7x6rjg4ntzqjhedm5b5zjqulugknhzr46ny2qbad.onion")