+++ crates/nostr-sdk/examples/bot.rs
@@ -12,7 +12,7 @@ async fn main() -> Result<()> {
 
     let keys = Keys::parse("nsec12kcgs78l06p30jz7z7h3n2x2cy99nw2z6zspjdp7qc206887mwvs95lnkx")?;
     let opts = Options::new().connection_timeout(Some(Duration::from_secs(10)));
-    let client = Client::with_opts(keys.clone(), opts);
+    let client = Client::builder().signer(keys.clone()).opts(opts).build();
 
     println!("Bot public key: {}", keys.public_key().to_bech32()?);