+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * pool: switch from async to sync message sending for `Relay` ([Yuki Kishimoto])
 * sdk: disable all default features ([Yuki Kishimoto])
 * sdk: set `Client::from_builder` as private ([Yuki Kishimoto])
+* sdk: deprecate `Client::with_opts` ([Yuki Kishimoto])
 
 ### Added